refactor: move state util fn to utils file
