+++ src/lib/toolsHandler.js
@@ -33,7 +33,10 @@ export default class ToolsHandler {
     const { zoom } = store.getState();
     const { left, top } = this._canvas.canvas.getBoundingClientRect();
 
-    return [(evt.pageX - left) / zoom, (evt.pageY - top) / zoom];
+    return this.getFixedCoords([
+      (evt.pageX - left) / zoom,
+      (evt.pageY - top) / zoom,
+    ]);
   }
 
   /* Cuando dibuja, en pointer move */
@@ -70,13 +73,11 @@ export default class ToolsHandler {
     this.setPrevAxis(axis);
     this._canvas.applySettings(this._toolSetting);
     this._toolState.ctx = this._canvas.context;
-    // Resuelve el problema de
-    this._toolState.ctx.translate(0.5, 0.5);
+    // this._toolState.ctx.translate(0.5, 0.5); No funciona como esperaba...
     this._points = [axis];
   }
 
   /**
-   * @deprecated El problema lo soluciona 'ctx.translate(0.5, 0.5);'
    * Sirve para evitar un bug al dibujar en cordenadas verticales u horizantales con linewidth par o impares.
    * Linewidth es impar, es mejor que la coord termine con .5.
    * Linewdith, es par, es mejor que sea un numero entero.
